+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+
+## 7.1.4
+
+* Typescript: Fix tsify and browserify compatibility
+* HistoryController: Remove UrlParser for window.URL
+
+
 ## 7.1.3
 
 * Typescript: Fix some type